chore(data) refactoring fee and blockState  methods
